# deploy/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Driftless — Issuer-Processor Core API
  description: >
    Public REST surface of Driftless: an immutable double-entry ledger plus a card-authorization
    saga that keeps money correct under failure. Every mutating endpoint is idempotent on the
    `Idempotency-Key` header (missing ⇒ 400, same key + different body ⇒ 409, replay ⇒ original 2xx).
    Money is integer **minor units** with an explicit ISO-4217 currency — no floating point ever
    crosses the boundary. Errors are RFC-7807 `application/problem+json` bodies carrying a stable,
    machine-readable `code`.


    This document is hand-authored and committed in-repo (`deploy/openapi/openapi.yaml`). springdoc
    `2.x` targets Spring 6 / Spring Boot 3 and is not yet compatible with the project's Spring Boot 4
    runtime, so it would break the build; the committed contract is served statically by the app at
    `/openapi.yaml` and rendered by Swagger UI at `/swagger-ui/index.html`.
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Local modular monolith (docker-compose / `app`)
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque, non-blank key making the mutating request safe to retry.
      schema:
        type: string
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ProblemDetail:
      description: RFC-7807 problem with a stable machine-readable `code`.
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
          description: >
            Stable error code clients branch on, e.g. VALIDATION_FAILED, MISSING_IDEMPOTENCY_KEY,
            IDEMPOTENCY_CONFLICT, ACCOUNT_NOT_FOUND, AUTHORIZATION_NOT_FOUND, TOKEN_NOT_FOUND,
            ILLEGAL_AUTHORIZATION_STATE, ILLEGAL_TOKEN_TRANSITION, BALANCE_INVARIANT_VIOLATION,
            CURRENCY_MISMATCH, DOMAIN_ERROR.
    OpenAccountRequest:
      type: object
      required: [currency, name]
      properties:
        type:
          type: string
          description: Double-entry account category; defaults to LIABILITY when omitted.
          enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
        name:
          type: string
          example: cardholder-alice
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        currency:
          type: string
        name:
          type: string
    FundAccountRequest:
      type: object
      required: [amountMinor, currency]
      properties:
        amountMinor:
          type: integer
          format: int64
          minimum: 1
          description: Amount to load, in minor units (strictly positive).
          example: 100000
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
    FundingResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        postedAt:
          type: string
          format: date-time
...
    StatementEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
        direction:
          type: string
          enum: [DEBIT, CREDIT]
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
        sequenceNo:
          type: integer
    StatementResponse:
      type: object
      properties:
        account:
          type: string
          format: uuid
        entries:
          type: array
          items:
            $ref: '#/components/schemas/StatementEntry'
        nextCursor:
          type: [string, "null"]
          description: Opaque keyset cursor for the next page; null at the end of the stream.
...
  responses:
    Problem:
      description: RFC-7807 problem response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
paths:
  /accounts:
    post:
      tags: [Accounts]
      summary: Open a ledger account
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAccountRequest'
      responses:
        '201':
          description: Account opened (replay returns the original).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /accounts/{id}:
    get:
      tags: [Accounts]
      summary: Get account details
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Account details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404': { $ref: '#/components/responses/Problem' }
  /accounts/{id}/funding:
    post:
      tags: [Accounts]
      summary: Fund (load) an account with a balanced ledger post
      description: >
        Loads funds onto the account with a balanced double-entry post — a DEBIT to the target and a
        CREDIT to a per-currency system funding source — keeping the global Σ(journal entries) == 0.
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FundAccountRequest'
      responses:
        '200':
          description: Funding transaction posted (replay returns the original).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingResponse'
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /accounts/{id}/statement:
    get:
      tags: [Accounts]
      summary: Keyset (seek) statement page
      parameters:
        - $ref: '#/components/parameters/PathId'
        - name: after
          in: query
          required: false
          description: Opaque cursor from a previous page's nextCursor; omit for the first page.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Page size, capped at 200 (default 50).
          schema:
            type: integer
            maximum: 200
      responses:
        '200':
          description: A page of entries plus the next cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementResponse'
        '404': { $ref: '#/components/responses/Problem' }
...